Finally got a render of the scene in a texture in the “same” scene. It’s supposed to be a circle, but it’s a square right now because the math is simpler. And it’s suppose to be magnifying, but it’s minifying (since it’s a sparse scene, and it’s easier to see).
Current “test” is creating a texture prim, and then, when we have a
single or left mouse tap, we replace the texture with another texture
which is the scene itself.

Still a few hardcoded bits hacked here and there.

screen conversions:
Working on converting the geometry of the prim to texture points. It’s
still shaky.

Rendering:
Callback after the pre-render (or rendering offscreen) so we have a
chance to grab the texture to set it for the actual render. Added
passing in a mode to the _GetDrawMode. we can pass back a solid or a
lined mode. Maybe depends on the shader?

Shaders:
The scene has a basic color shader, and a texture shader. At some
point, we’ll have a lighting shader. I’d like a tessellation shader too.