Checkpoint. Did a lot of rework to get the texture “magnification” working. It required some rethinking of render passes.
Note: The frequent crashes on quit are still there

I think I simplified the different coordinate spaces to some degree.
 o World Space: We have models that live in world space
 o NDC: Screen space prims are orthographic 3D
 o Coordinate Screen: NDC in the screen without Z
 o Screen: Bottom Left (0,0) to Top right (1,1)
 o Texture: Top Left (0,0) to Bottom Right (1,1)

Reduced the number of conversions between these spaces to just what we
need (hopefully).

For now, using a square prim for the loupe since that simplifies the
math and makes it easier to debug.

Moved frustum computations to a FrustumCompute class. I think adds a
little efficiency as well as making it easier to test.

Scene Rendering:
Instead of keeping track of the stencil prims, we separate out the
screen space prims and regular scene prims. When we draw the stencil
prims, we set the stencil mask, and unset it after. Seems like it
should work, although we only have 1 screen space prim at the moment.

On scene render, we set the screen space camera values on the shader
uniforms, and draw those prims, then set the regular camera values on
the shader, and then draw those prims.

Two tests that are barely written for the frustum compute and frustum.