Two fixes, and some cleanup
Cleanup:
Tried to standardize coordinate system conversions. Much closer now.

Prim wrapped a couple calls to Scene. Got rid of that and added a
_GetScene(). This saves an indirection.

Two main fixes:
 o The big fix was reinitialization prims, because initializeGL() is
called before resize(). This resulted in creating the geometry with
different dimensions than we display.
 o _GetHeightFromDistance() was using the whole FOV. This is
incorrect. We need to halve it. Then we have a right angle’d triangle.